@@ -15,6 +15,8 @@
  * the License.
  */
 
+using System.Reflection;
+
 namespace Aerospike.Client
 {
 	/// <summary>
@@ -42,6 +44,8 @@ namespace Aerospike.Client
 		//-------------------------------------------------------
 
 		private readonly new AsyncCluster cluster;
+
+		protected internal new AsyncClientPolicy clientPolicy;
 
 		//-------------------------------------------------------
 		// Constructors
@@ -118,14 +122,38 @@ namespace Aerospike.Client
 		/// <param name="hosts">array of potential hosts to seed the cluster</param>
 		/// <exception cref="AerospikeException">if all host connections fail</exception>
 		public AsyncClient(AsyncClientPolicy policy, params Host[] hosts)
-			: base (policy)
+			: base(policy)
 		{
-			if (policy == null)
+			policy ??= new AsyncClientPolicy();
+			clientPolicy = policy;
+
+			string configPath = YamlConfigProvider.GetConfigPath();
+			if (configPath != null)
 			{
-				policy = new AsyncClientPolicy();
+				configProvider = YamlConfigProvider.CreateConfigProvider(configPath, this);
 			}
-			this.cluster = new AsyncCluster(policy, hosts);
+			else
+			{
+				configProvider = null;
+			}
+
+			if (configProvider != null)
+			{
+				policy = new AsyncClientPolicy(policy, configProvider);
+				MergeDefaultPoliciesWithConfig();
+			}
+			
+			version = Assembly.GetExecutingAssembly().GetName().Version.ToString();
+			version ??= "development";
+
+			cluster = new AsyncCluster(this, policy, configPath, hosts);
 			base.cluster = this.cluster;
+			base.clientPolicy = this.clientPolicy;
+		}
+
+		protected internal override AsyncClientPolicy GetClientPolicy()
+		{
+			return clientPolicy;
 		}
 
 		//-------------------------------------------------------
@@ -162,7 +190,7 @@ namespace Aerospike.Client
 		public void Commit(CommitListener listener, Txn txn)
 		{
 			AsyncTxnRoll atr = new(
-				cluster, txnVerifyPolicyDefault, txnRollPolicyDefault, txn
+				cluster, mergedTxnVerifyPolicyDefault, mergedTxnRollPolicyDefault, txn
 				);
 
 			switch (txn.State)
@@ -213,7 +241,7 @@ namespace Aerospike.Client
 		/// <param name="txn">transaction</param>
 		public void Abort(AbortListener listener, Txn txn)
 		{
-			AsyncTxnRoll atr = new(cluster, null, txnRollPolicyDefault, txn);
+			AsyncTxnRoll atr = new(cluster, null, mergedTxnRollPolicyDefault, txn);
 			
 			switch (txn.State)
 			{
@@ -274,8 +302,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncWrite async = new AsyncWrite(cluster, policy, listener, key, bins, Operation.Type.WRITE);
 			AsyncTxnMonitor.Execute(cluster, policy, async);
 		}
@@ -324,8 +357,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncWrite async = new AsyncWrite(cluster, policy, listener, key, bins, Operation.Type.APPEND);
 			AsyncTxnMonitor.Execute(cluster, policy, async);
 		}
@@ -370,8 +408,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncWrite async = new AsyncWrite(cluster, policy, listener, key, bins, Operation.Type.PREPEND);
 			AsyncTxnMonitor.Execute(cluster, policy, async);
 		}
@@ -418,8 +461,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncWrite async = new AsyncWrite(cluster, policy, listener, key, bins, Operation.Type.ADD);
 			AsyncTxnMonitor.Execute(cluster, policy, async);
 		}
@@ -456,8 +504,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncDelete async = new AsyncDelete(cluster, policy, key, listener);
 			AsyncTxnMonitor.Execute(cluster, policy, async);
 		}
@@ -504,12 +557,20 @@ namespace Aerospike.Client
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
+			}
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
 			}
 
 			if (deletePolicy == null)
 			{
-				deletePolicy = batchDeletePolicyDefault;
+				deletePolicy = mergedBatchDeletePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				deletePolicy = new BatchDeletePolicy(deletePolicy, configProvider);
 			}
 
 			BatchAttr attr = new BatchAttr();
@@ -545,12 +606,20 @@ namespace Aerospike.Client
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
+			}
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
 			}
 
 			if (deletePolicy == null)
 			{
-				deletePolicy = batchDeletePolicyDefault;
+				deletePolicy = mergedBatchDeletePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				deletePolicy = new BatchDeletePolicy(deletePolicy, configProvider);
 			}
 
 			BatchAttr attr = new BatchAttr();
@@ -594,8 +663,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncTouch async = new AsyncTouch(cluster, policy, listener, key);
 			AsyncTxnMonitor.Execute(cluster, policy, async);
 		}
@@ -634,8 +708,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncTouch async = new(cluster, policy, listener, key);
 			AsyncTxnMonitor.Execute(cluster, policy, async);
 		}
@@ -672,7 +751,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = readPolicyDefault;
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -709,13 +792,19 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				listener.OnSuccess(keys, new bool[0]);
+				listener.OnSuccess(keys, []);
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			AsyncBatchExistsArrayExecutor executor = new(cluster, policy, keys, listener);
@@ -738,10 +827,16 @@ namespace Aerospike.Client
 				listener.OnSuccess();
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			AsyncBatchExistsSequenceExecutor executor = new(cluster, policy, keys, listener);
@@ -780,7 +875,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = readPolicyDefault;
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -819,7 +918,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = readPolicyDefault;
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -856,7 +959,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = readPolicyDefault;
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -909,10 +1016,16 @@ namespace Aerospike.Client
 				listener.OnSuccess(records);
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(records);
 
 			AsyncBatchReadListExecutor executor = new(cluster, policy, listener, records);
@@ -940,10 +1053,16 @@ namespace Aerospike.Client
 				listener.OnSuccess();
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(records);
 
 			AsyncBatchReadSequenceExecutor executor = new(cluster, policy, listener, records);
@@ -984,13 +1103,19 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				listener.OnSuccess(keys, new Record[0]);
+				listener.OnSuccess(keys, []);
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			AsyncBatchGetArrayExecutor executor = new(cluster, policy, listener, keys, null, null, Command.INFO1_READ | Command.INFO1_GET_ALL, false);
@@ -1016,10 +1141,16 @@ namespace Aerospike.Client
 				listener.OnSuccess();
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			AsyncBatchGetSequenceExecutor executor = new(cluster, policy, listener, keys, null, null, Command.INFO1_READ | Command.INFO1_GET_ALL, false);
@@ -1062,13 +1193,19 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				listener.OnSuccess(keys, new Record[0]);
+				listener.OnSuccess(keys, []);
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			AsyncBatchGetArrayExecutor executor = new(cluster, policy, listener, keys, binNames, null, Command.INFO1_READ, false);
@@ -1095,10 +1232,16 @@ namespace Aerospike.Client
 				listener.OnSuccess();
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			int readAttr = (binNames == null || binNames.Length == 0)?
@@ -1145,13 +1288,17 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				listener.OnSuccess(keys, new Record[0]);
+				listener.OnSuccess(keys, []);
 				return;
 			}
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(keys);
@@ -1184,7 +1331,11 @@ namespace Aerospike.Client
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(keys);
@@ -1227,13 +1378,19 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				listener.OnSuccess(keys, new Record[0]);
+				listener.OnSuccess(keys, []);
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			AsyncBatchGetArrayExecutor executor = new(cluster, policy, listener, keys, null, null, Command.INFO1_READ | Command.INFO1_NOBINDATA, false);
@@ -1259,10 +1416,16 @@ namespace Aerospike.Client
 				listener.OnSuccess();
 				return;
 			}
+
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
 			policy.Txn?.PrepareRead(keys);
 
 			AsyncBatchGetSequenceExecutor executor = new(cluster, policy, listener, keys, null, null, Command.INFO1_READ | Command.INFO1_NOBINDATA, false);
@@ -1314,8 +1477,12 @@ namespace Aerospike.Client
 		/// <exception cref="AerospikeException">if queue is full</exception>
 		public void Operate(WritePolicy policy, RecordListener listener, Key key, params Operation[] ops)
 		{
-			OperateArgs args = new OperateArgs(policy, writePolicyDefault, operatePolicyReadDefault, ops);
+			OperateArgs args = new OperateArgs(policy, mergedWritePolicyDefault, mergedOperatePolicyReadDefault, ops);
 			policy = args.writePolicy;
+			if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
 
 			if (args.hasWrite)
 			{
@@ -1380,9 +1547,14 @@ namespace Aerospike.Client
 
 			if (policy == null)
 			{
-				policy = batchParentPolicyWriteDefault;
+				policy = mergedBatchPolicyDefault;
 			}
-			AsyncBatchOperateListExecutor executor = new(cluster, policy, listener, records);
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
+			AsyncBatchOperateListExecutor executor = new(cluster, policy, listener, records, configProvider);
 			AsyncTxnMonitor.ExecuteBatch(policy, executor, records);
 		}
 
@@ -1415,9 +1587,14 @@ namespace Aerospike.Client
 
 			if (policy == null)
 			{
-				policy = batchParentPolicyWriteDefault;
+				policy = mergedBatchPolicyDefault;
 			}
-			AsyncBatchOperateSequenceExecutor executor = new(cluster, policy, listener, records);
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
+			}
+
+			AsyncBatchOperateSequenceExecutor executor = new(cluster, policy, listener, records, configProvider);
 			AsyncTxnMonitor.ExecuteBatch(policy, executor, records);
 		}
 
@@ -1473,15 +1650,26 @@ namespace Aerospike.Client
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
 			}
-
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
+			}
 			if (writePolicy == null)
 			{
-				writePolicy = batchWritePolicyDefault;
+				writePolicy = mergedBatchWritePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				writePolicy = new BatchWritePolicy(writePolicy, configProvider);
 			}
 
-			BatchAttr attr = new BatchAttr(batchPolicy, writePolicy, ops);
+			BatchAttr attr = new(batchPolicy, writePolicy, ops);
+			if (attr.hasWrite && configProvider != null)
+			{
+				batchPolicy.GraftBatchWriteConfig(configProvider);
+			}
 			AsyncBatchOperateRecordArrayExecutor executor = new(cluster, batchPolicy, listener, keys, ops, attr);
 			AsyncTxnMonitor.ExecuteBatch(batchPolicy, executor, keys);
 		}
@@ -1517,15 +1705,26 @@ namespace Aerospike.Client
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
 			}
-
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
+			}
 			if (writePolicy == null)
 			{
-				writePolicy = batchWritePolicyDefault;
+				writePolicy = mergedBatchWritePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				writePolicy = new BatchWritePolicy(writePolicy, configProvider);
 			}
 
-			BatchAttr attr = new BatchAttr(batchPolicy, writePolicy, ops);
+			BatchAttr attr = new(batchPolicy, writePolicy, ops);
+			if (attr.hasWrite && configProvider != null)
+			{
+				batchPolicy.GraftBatchWriteConfig(configProvider);
+			}
 			AsyncBatchOperateRecordSequenceExecutor executor = new(cluster, batchPolicy, listener, keys, ops, attr);
 			AsyncTxnMonitor.ExecuteBatch(batchPolicy, executor, keys);
 		}
@@ -1553,7 +1752,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = scanPolicyDefault;
+				policy = mergedScanPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new ScanPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -1581,7 +1784,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = scanPolicyDefault;
+				policy = mergedScanPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new ScanPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -1634,8 +1841,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			AsyncExecute command = new AsyncExecute(cluster, policy, listener, key, packageName, functionName, functionArgs);
 			AsyncTxnMonitor.Execute(cluster, policy, command);
 		}
@@ -1690,12 +1902,19 @@ namespace Aerospike.Client
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
 			}
-
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
+			}
 			if (udfPolicy == null)
 			{
-				udfPolicy = batchUDFPolicyDefault;
+				udfPolicy = mergedBatchUDFPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				udfPolicy = new BatchUDFPolicy(udfPolicy, configProvider);
 			}
 
 			byte[] argBytes = Packer.Pack(functionArgs);
@@ -1738,12 +1957,19 @@ namespace Aerospike.Client
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
 			}
-
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
+			}
 			if (udfPolicy == null)
 			{
-				udfPolicy = batchUDFPolicyDefault;
+				udfPolicy = mergedBatchUDFPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				udfPolicy = new BatchUDFPolicy(udfPolicy, configProvider);
 			}
 
 			byte[] argBytes = Packer.Pack(functionArgs);
@@ -1779,7 +2005,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = queryPolicyDefault;
+				policy = mergedQueryPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new QueryPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -1825,7 +2055,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = queryPolicyDefault;
+				policy = mergedQueryPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new QueryPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
