@@ -46,64 +46,79 @@ namespace Aerospike.Client
 
 		protected internal Cluster cluster;
 
+		protected internal string version;
+
+		protected internal ClientPolicy clientPolicy;
+
 		/// <summary>
 		/// Default read policy that is used when read command policy is null.
 		/// </summary>
 		protected Policy readPolicyDefault;
+		internal Policy mergedReadPolicyDefault;
 
 		/// <summary>
 		/// Default write policy that is used when write command policy is null.
 		/// </summary>
 		protected WritePolicy writePolicyDefault;
+		internal WritePolicy mergedWritePolicyDefault;
 
 		/// <summary>
 		/// Default scan policy that is used when scan command policy is null.
 		/// </summary>
 		protected ScanPolicy scanPolicyDefault;
+		internal ScanPolicy mergedScanPolicyDefault;
 
 		/// <summary>
 		/// Default query policy that is used when query command policy is null.
 		/// </summary>
 		protected QueryPolicy queryPolicyDefault;
+		internal QueryPolicy mergedQueryPolicyDefault;
 
 		/// <summary>
 		/// Default parent policy used in batch read commands. Parent policy fields
 		/// include socketTimeout, totalTimeout, maxRetries, etc...
 		/// </summary>
 		protected BatchPolicy batchPolicyDefault;
+		internal BatchPolicy mergedBatchPolicyDefault;
 
 		/// <summary>
 		/// Default parent policy used in batch write commands. Parent policy fields
 		/// include socketTimeout, totalTimeout, maxRetries, etc...
 		/// </summary>
 		protected BatchPolicy batchParentPolicyWriteDefault;
+		internal BatchPolicy mergedBatchParentPolicyWriteDefault;
 
 		/// <summary>
 		/// Default write policy used in batch operate commands.
 		/// Write policy fields include generation, expiration, durableDelete, etc...
 		/// </summary>
 		protected BatchWritePolicy batchWritePolicyDefault;
+		internal BatchWritePolicy mergedBatchWritePolicyDefault;
 
 		/// <summary>
 		/// Default delete policy used in batch delete commands.
 		/// </summary>
 		protected BatchDeletePolicy batchDeletePolicyDefault;
+		internal BatchDeletePolicy mergedBatchDeletePolicyDefault;
 
 		/// <summary>
 		/// Default user defined function policy used in batch UDF execute commands.
 		/// </summary>
 		protected BatchUDFPolicy batchUDFPolicyDefault;
+		internal BatchUDFPolicy mergedBatchUDFPolicyDefault;
 
 		/// <summary>
 		/// Default transaction policy when verifying record versions in a batch on a commit.
 		/// </summary>
 		protected TxnVerifyPolicy txnVerifyPolicyDefault;
+		internal TxnVerifyPolicy mergedTxnVerifyPolicyDefault;
 
 		/// <summary>
 		/// Default transaction policy when rolling the transaction records forward (commit)
 		/// or back(abort) in a batch.
 		/// </summary>
 		protected TxnRollPolicy txnRollPolicyDefault;
+		internal TxnRollPolicy mergedTxnRollPolicyDefault;
 
 		/// <summary>
 		/// Default info policy that is used when info command policy is null.
@@ -111,6 +126,9 @@ namespace Aerospike.Client
 		protected InfoPolicy infoPolicyDefault;
 
 		protected WritePolicy operatePolicyReadDefault;
+		internal WritePolicy mergedOperatePolicyReadDefault;
+
+		internal protected IConfigProvider configProvider;
 
 		//-------------------------------------------------------
 		// Constructors
@@ -188,10 +206,8 @@ namespace Aerospike.Client
 		/// <exception cref="AerospikeException">if all host connections fail</exception>
 		public AerospikeClient(ClientPolicy policy, params Host[] hosts)
 		{
-			if (policy == null)
-			{
-				policy = new ClientPolicy();
-			}
+			policy ??= new ClientPolicy();
+			clientPolicy = policy;
 			this.readPolicyDefault = policy.readPolicyDefault;
 			this.writePolicyDefault = policy.writePolicyDefault;
 			this.scanPolicyDefault = policy.scanPolicyDefault;
@@ -206,7 +222,40 @@ namespace Aerospike.Client
 			this.infoPolicyDefault = policy.infoPolicyDefault;
 			this.operatePolicyReadDefault = new WritePolicy(this.readPolicyDefault);
 
-			cluster = new Cluster(policy, hosts);
+			mergedReadPolicyDefault = this.readPolicyDefault;
+			mergedWritePolicyDefault = this.writePolicyDefault;
+			mergedScanPolicyDefault = this.scanPolicyDefault;
+			mergedQueryPolicyDefault = this.queryPolicyDefault;
+			mergedBatchPolicyDefault = this.batchPolicyDefault;
+			mergedBatchParentPolicyWriteDefault = this.batchParentPolicyWriteDefault;
+			mergedBatchWritePolicyDefault = this.batchWritePolicyDefault;
+			mergedBatchDeletePolicyDefault = this.batchDeletePolicyDefault;
+			mergedBatchUDFPolicyDefault = this.batchUDFPolicyDefault;
+			mergedTxnVerifyPolicyDefault = this.txnVerifyPolicyDefault;
+			mergedTxnRollPolicyDefault = this.txnRollPolicyDefault;
+			mergedOperatePolicyReadDefault = this.operatePolicyReadDefault;
+
+			string configPath = YamlConfigProvider.GetConfigPath();
+			if (configPath != null)
+			{
+				configProvider = YamlConfigProvider.CreateConfigProvider(configPath, this);
+			}
+			else
+			{
+				configProvider = null;
+			}
+
+			if (configProvider != null)
+			{
+				policy = new ClientPolicy(policy, configProvider);
+				MergeDefaultPoliciesWithConfig();
+			}
+
+			version = Assembly.GetExecutingAssembly().GetName().Version.ToString();
+			version ??= "development";
+
+			cluster = new Cluster(this, policy, configPath, hosts);
+			cluster.UpdateClusterConfig(true);
 			cluster.StartTendThread(policy);
 		}
 
@@ -248,6 +297,24 @@ namespace Aerospike.Client
 				this.infoPolicyDefault = new InfoPolicy();
 			}
 			this.operatePolicyReadDefault = new WritePolicy(this.readPolicyDefault);
+
+			mergedReadPolicyDefault = this.readPolicyDefault;
+			mergedWritePolicyDefault = this.writePolicyDefault;
+			mergedScanPolicyDefault = this.scanPolicyDefault;
+			mergedQueryPolicyDefault = this.queryPolicyDefault;
+			mergedBatchPolicyDefault = this.batchPolicyDefault;
+			mergedBatchParentPolicyWriteDefault = this.batchParentPolicyWriteDefault;
+			mergedBatchWritePolicyDefault = this.batchWritePolicyDefault;
+			mergedBatchDeletePolicyDefault = this.batchDeletePolicyDefault;
+			mergedBatchUDFPolicyDefault = this.batchUDFPolicyDefault;
+			mergedTxnVerifyPolicyDefault = this.txnVerifyPolicyDefault;
+			mergedTxnRollPolicyDefault = this.txnRollPolicyDefault;
+			mergedOperatePolicyReadDefault = this.operatePolicyReadDefault;
+		}
+
+		protected virtual internal ClientPolicy GetClientPolicy()
+		{
+			return clientPolicy;
 		}
 
 		//-------------------------------------------------------
@@ -390,6 +457,27 @@ namespace Aerospike.Client
 			set { txnRollPolicyDefault = value; }
 		}
 
+		/// <summary>
+		/// Merge the default policies with the config properties.  This should be done at init and every time
+		/// the config is updated.
+		/// </summary>
+		public void MergeDefaultPoliciesWithConfig()
+		{
+			mergedReadPolicyDefault = new Policy(readPolicyDefault, configProvider);
+			mergedWritePolicyDefault = new WritePolicy(writePolicyDefault, configProvider);
+			mergedScanPolicyDefault = new ScanPolicy(scanPolicyDefault, configProvider);
+			mergedQueryPolicyDefault = new QueryPolicy(queryPolicyDefault, configProvider);
+			mergedBatchPolicyDefault = new BatchPolicy(batchPolicyDefault, configProvider);
+			mergedBatchParentPolicyWriteDefault = new BatchPolicy(batchParentPolicyWriteDefault, configProvider);
+			mergedBatchParentPolicyWriteDefault.GraftBatchWriteConfig(configProvider);
+			mergedBatchWritePolicyDefault = new BatchWritePolicy(batchWritePolicyDefault, configProvider);
+			mergedBatchDeletePolicyDefault = new BatchDeletePolicy(batchDeletePolicyDefault, configProvider);
+			mergedBatchUDFPolicyDefault = new BatchUDFPolicy(batchUDFPolicyDefault, configProvider);
+			mergedTxnVerifyPolicyDefault = new TxnVerifyPolicy(txnVerifyPolicyDefault, configProvider);
+			mergedTxnRollPolicyDefault = new TxnRollPolicy(txnRollPolicyDefault, configProvider);
+			mergedOperatePolicyReadDefault = new WritePolicy(operatePolicyReadDefault, configProvider);
+		}
+
 		//-------------------------------------------------------
 		// Cluster Connection Management
 		//-------------------------------------------------------
@@ -506,11 +594,11 @@ namespace Aerospike.Client
 			{
 				default:
 				case Txn.TxnState.OPEN:
-					tr.Verify(txnVerifyPolicyDefault, txnRollPolicyDefault);
-					return tr.Commit(txnRollPolicyDefault);
+					tr.Verify(mergedTxnVerifyPolicyDefault, mergedTxnRollPolicyDefault);
+					return tr.Commit(mergedTxnRollPolicyDefault);
 				
 				case Txn.TxnState.VERIFIED:
-					return tr.Commit(txnRollPolicyDefault);
+					return tr.Commit(mergedTxnRollPolicyDefault);
 				
 				case Txn.TxnState.COMMITTED:
 					return CommitStatus.CommitStatusType.ALREADY_COMMITTED;
@@ -537,7 +625,7 @@ namespace Aerospike.Client
 				default:
 				case Txn.TxnState.OPEN:
 				case Txn.TxnState.VERIFIED:
-					return tr.Abort(txnRollPolicyDefault);
+					return tr.Abort(mergedTxnRollPolicyDefault);
 
 				case Txn.TxnState.COMMITTED:
 					throw new AerospikeException(ResultCode.TXN_ALREADY_COMMITTED, "Transaction already committed");
@@ -564,7 +652,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			if (policy.Txn != null)
@@ -594,7 +686,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			if (policy.Txn != null)
@@ -620,7 +716,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			if (policy.Txn != null)
@@ -649,7 +749,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			if (policy.Txn != null)
@@ -677,7 +781,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if(configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			if (policy.Txn != null)
@@ -705,17 +813,25 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				return new BatchResults(new BatchRecord[0], true);
+				return new BatchResults([], true);
 			}
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
+			} 
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
 			}
 
 			if (deletePolicy == null)
 			{
-				deletePolicy = batchDeletePolicyDefault;
+				deletePolicy = mergedBatchDeletePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				deletePolicy = new BatchDeletePolicy(deletePolicy, configProvider);
 			}
 
 			if (batchPolicy.Txn != null)
@@ -777,10 +893,7 @@ namespace Aerospike.Client
 		/// </param>
 		public void Truncate(InfoPolicy policy, string ns, string set, DateTime? beforeLastUpdate)
 		{
-			if (policy == null)
-			{
-				policy = infoPolicyDefault;
-			}
+			policy ??= infoPolicyDefault;
 
 			// Send truncate command to one node. That node will distribute the command to other nodes.
 			Node node = cluster.GetRandomNode();
@@ -831,9 +944,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
-			
+			if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			if (policy.Txn != null)
 			{
 				TxnMonitor.AddKey(cluster, policy, key);
@@ -856,8 +973,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
+
 			if (policy.Txn != null)
 			{
 				TxnMonitor.AddKey(cluster, policy, key);
@@ -885,7 +1007,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = readPolicyDefault;
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -906,12 +1032,16 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				return new bool[0];
+				return [];
 			}
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(keys);
@@ -964,8 +1094,12 @@ namespace Aerospike.Client
 		public Record Get(Policy policy, Key key)
 		{
 			if (policy == null)
+			{ 
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
 			{
-				policy = readPolicyDefault;
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -988,7 +1122,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = readPolicyDefault;
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -1010,7 +1148,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = readPolicyDefault;
+				policy = mergedReadPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(key.ns);
@@ -1044,7 +1186,11 @@ namespace Aerospike.Client
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(records);
@@ -1074,12 +1220,16 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				return new Record[0];
+				return [];
 			}
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(keys);
@@ -1130,12 +1280,16 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				return new Record[0];
+				return [];
 			}
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(keys);
@@ -1191,7 +1345,11 @@ namespace Aerospike.Client
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(keys);
@@ -1241,12 +1399,16 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				return new Record[0];
+				return [];
 			}
 
 			if (policy == null)
 			{
-				policy = batchPolicyDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			policy.Txn?.PrepareRead(keys);
@@ -1353,8 +1515,12 @@ namespace Aerospike.Client
 		/// <exception cref="AerospikeException">if command fails</exception>
 		public Record Operate(WritePolicy policy, Key key, params Operation[] operations)
 		{
-			OperateArgs args = new OperateArgs(policy, writePolicyDefault, operatePolicyReadDefault, operations);
+			OperateArgs args = new OperateArgs(policy, mergedWritePolicyDefault, mergedOperatePolicyReadDefault, operations);
 			policy = args.writePolicy;
+			if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
+			}
 
 			if (args.hasWrite)
 			{
@@ -1409,7 +1575,11 @@ namespace Aerospike.Client
 
 			if (policy == null)
 			{
-				policy = batchParentPolicyWriteDefault;
+				policy = mergedBatchPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new BatchPolicy(policy, configProvider);
 			}
 
 			if (policy.Txn != null)
@@ -1424,7 +1594,7 @@ namespace Aerospike.Client
 
 			foreach (BatchNode batchNode in batchNodes)
 			{
-				commands[count++] = new BatchOperateListCommand(cluster, batchNode, policy, records, status);
+				commands[count++] = new BatchOperateListCommand(cluster, batchNode, policy, records, status, configProvider);
 			}
 			BatchExecutor.Execute(cluster, policy, commands, status);
 			return status.GetStatus();
@@ -1455,20 +1625,31 @@ namespace Aerospike.Client
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
 			}
-
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
+			}
 			if (writePolicy == null)
 			{
-				writePolicy = batchWritePolicyDefault;
+				writePolicy = mergedBatchWritePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				writePolicy = new BatchWritePolicy(writePolicy, configProvider);
 			}
 
 			if (batchPolicy.Txn != null)
 			{
 				TxnMonitor.AddKeys(cluster, batchPolicy, keys);
-			}
+			} 
 
-			BatchAttr attr = new BatchAttr(batchPolicy, writePolicy, ops);
+			BatchAttr attr = new(batchPolicy, writePolicy, ops);
+			if (attr.hasWrite && configProvider != null)
+			{
+				batchPolicy.GraftBatchWriteConfig(configProvider);
+			}
 			BatchRecord[] records = new BatchRecord[keys.Length];
 
 			for (int i = 0; i < keys.Length; i++)
@@ -1522,7 +1703,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = scanPolicyDefault;
+				policy = mergedScanPolicyDefault;
+			}
+			else if (configProvider != null)
+			{ 
+				policy = new ScanPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -1573,7 +1758,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = scanPolicyDefault;
+				policy = mergedScanPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new ScanPolicy(policy, configProvider);
 			}
 
 			PartitionTracker tracker = new PartitionTracker(policy, node);
@@ -1598,7 +1787,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = scanPolicyDefault;
+				policy = mergedScanPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new ScanPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -1624,9 +1817,14 @@ namespace Aerospike.Client
 		public RegisterTask Register(Policy policy, string clientPath, string serverPath, Language language)
 		{
 			if (policy == null)
-			{
-				policy = writePolicyDefault;
+			{ 
+				policy = mergedWritePolicyDefault;
 			}
+			else if (configProvider != null) 
+			{
+				policy = new Policy(policy, configProvider);
+			}
+
 			string content = Util.ReadFileEncodeBase64(clientPath);
 			return RegisterCommand.Register(cluster, policy, content, serverPath, language);
 		}
@@ -1647,8 +1845,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
+			}
+
 			string content;
 			using (Stream stream = resourceAssembly.GetManifestResourceStream(resourcePath))
 			{
@@ -1692,8 +1895,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
+			}
+
 			byte[] bytes = ByteUtil.StringToUtf8(code);
 			string content = Convert.ToBase64String(bytes);
 			return RegisterCommand.Register(cluster, policy, content, serverPath, language);
@@ -1707,10 +1915,8 @@ namespace Aerospike.Client
 		/// <exception cref="AerospikeException">if remove fails</exception>
 		public void RemoveUdf(InfoPolicy policy, string serverPath)
 		{
-			if (policy == null)
-			{
-				policy = infoPolicyDefault;
-			}
+			policy ??= infoPolicyDefault;
+
 			// Send UDF command to one node. That node will distribute the UDF command to other nodes.
 			string command = "udf-remove:filename=" + serverPath;
 			Node node = cluster.GetRandomNode();
@@ -1747,7 +1953,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			if (policy.Txn != null)
@@ -1801,17 +2011,24 @@ namespace Aerospike.Client
 		{
 			if (keys.Length == 0)
 			{
-				return new BatchResults(new BatchRecord[0], true);
+				return new BatchResults([], true);
 			}
 
 			if (batchPolicy == null)
 			{
-				batchPolicy = batchParentPolicyWriteDefault;
+				batchPolicy = mergedBatchParentPolicyWriteDefault;
 			}
-
+			else if (configProvider != null)
+			{
+				batchPolicy = new BatchPolicy(batchPolicy, configProvider);
+			}
 			if (udfPolicy == null)
 			{
-				udfPolicy = batchUDFPolicyDefault;
+				udfPolicy = mergedBatchUDFPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				udfPolicy = new BatchUDFPolicy(udfPolicy, configProvider);
 			}
 
 			if (batchPolicy.Txn != null)
@@ -1873,8 +2090,12 @@ namespace Aerospike.Client
 		public ExecuteTask Execute(WritePolicy policy, Statement statement, string packageName, string functionName, params Value[] functionArgs)
 		{
 			if (policy == null)
+			{ 
+				policy = mergedWritePolicyDefault;
+			}
+			else if (configProvider != null)
 			{
-				policy = writePolicyDefault;
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			statement.PackageName = packageName;
@@ -1912,7 +2133,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new WritePolicy(policy, configProvider);
 			}
 
 			if (operations.Length > 0)
@@ -1969,7 +2194,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = queryPolicyDefault;
+				policy = mergedQueryPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new QueryPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -2008,7 +2237,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = queryPolicyDefault;
+				policy = mergedQueryPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new QueryPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -2058,7 +2291,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = queryPolicyDefault;
+				policy = mergedQueryPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new QueryPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -2095,7 +2332,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = queryPolicyDefault;
+				policy = mergedQueryPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new QueryPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -2184,7 +2425,11 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = queryPolicyDefault;
+				policy = mergedQueryPolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new QueryPolicy(policy, configProvider);
 			}
 
 			Node[] nodes = cluster.ValidateNodes();
@@ -2250,9 +2495,56 @@ namespace Aerospike.Client
 			params CTX[] ctx
 		)
 		{
+			return CreateIndex(policy, ns, setName, indexName, binName, indexType, indexCollectionType, null, ctx);
+		}
+
+		/// <summary>
+		/// Create complex secondary index on bins containing collections.
+		/// This asynchronous server call will return before command is complete.
+		/// The user can optionally wait for command completion by using the returned
+		/// IndexTask instance.
+		/// </summary>
+		/// <param name="policy">generic configuration parameters, pass in null for defaults</param>
+		/// <param name="ns">namespace - equivalent to database name</param>
+		/// <param name="setName">optional set name - equivalent to database table</param>
+		/// <param name="indexName">name of secondary index</param>
+		/// <param name="indexType">underlying data type of secondary index</param>
+		/// <param name="indexCollectionType">index collection type</param>
+		/// <param name="indexExpression">the expression to be indexed</param>
+		/// <exception cref="AerospikeException">if index create fails</exception>
+		public IndexTask CreateIndex
+		(
+			Policy policy,
+			string ns,
+			string setName,
+			string indexName,
+			IndexType indexType,
+			IndexCollectionType indexCollectionType,
+			Expression indexExpression
+		)
+		{
+			return CreateIndex(policy, ns, setName, indexName, null, indexType, indexCollectionType, indexExpression, null);
+		}
+
+		private IndexTask CreateIndex(
+			Policy policy,
+			string ns,
+			string setName,
+			string indexName,
+			string binName,
+			IndexType indexType,
+			IndexCollectionType indexCollectionType,
+			Expression indexExpression,
+			params CTX[] ctx
+		)
+		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
+			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
 			}
 
 			StringBuilder sb = new StringBuilder(1024);
@@ -2268,25 +2560,42 @@ namespace Aerospike.Client
 			sb.Append(";indexname=");
 			sb.Append(indexName);
 
-			if (ctx != null && ctx.Length > 0)
+			if (indexExpression != null)
 			{
-				byte[] bytes = PackUtil.Pack(ctx);
-				string base64 = Convert.ToBase64String(bytes);
+				sb.Append(";exp=");
+				sb.Append(indexExpression.GetBase64());
 
-				sb.Append(";context=");
-				sb.Append(base64);
+				if (indexCollectionType != IndexCollectionType.DEFAULT)
+				{
+					sb.Append(";indextype=");
+					sb.Append(indexCollectionType);
+				}
+
+				sb.Append(";type=");
+				sb.Append(indexType);
 			}
-
-			if (indexCollectionType != IndexCollectionType.DEFAULT)
+			else
 			{
-				sb.Append(";indextype=");
-				sb.Append(indexCollectionType);
-			}
+				if (ctx != null && ctx.Length > 0)
+				{
+					byte[] bytes = PackUtil.Pack(ctx);
+					string base64 = Convert.ToBase64String(bytes);
 
-			sb.Append(";indexdata=");
-			sb.Append(binName);
-			sb.Append(',');
-			sb.Append(indexType);
+					sb.Append(";context=");
+					sb.Append(base64);
+				}
+
+				if (indexCollectionType != IndexCollectionType.DEFAULT)
+				{
+					sb.Append(";indextype=");
+					sb.Append(indexCollectionType);
+				}
+
+				sb.Append(";indexdata=");
+				sb.Append(binName);
+				sb.Append(',');
+				sb.Append(indexType);
+			}
 
 			// Send index command to one node. That node will distribute the command to other nodes.
 			String response = SendInfoCommand(policy, sb.ToString());
@@ -2316,8 +2625,13 @@ namespace Aerospike.Client
 		{
 			if (policy == null)
 			{
-				policy = writePolicyDefault;
+				policy = mergedWritePolicyDefault;
 			}
+			else if (configProvider != null)
+			{
+				policy = new Policy(policy, configProvider);
+			}
+
 			StringBuilder sb = new StringBuilder(500);
 			sb.Append("sindex-delete:ns=");
 			sb.Append(ns);
@@ -2357,10 +2671,7 @@ namespace Aerospike.Client
 		/// <exception cref="AerospikeException">if command fails</exception>
 		public void SetXDRFilter(InfoPolicy policy, string datacenter, string ns, Expression filter)
 		{
-			if (policy == null)
-			{
-				policy = infoPolicyDefault;
-			}
+			policy ??= infoPolicyDefault;
 
 			// Send XDR command to one node. That node will distribute the XDR command to other nodes.
 			string command = "xdr-set-filter:dc=" + datacenter + ";namespace=" + ns + ";exp=" + filter.GetBase64();
@@ -2395,6 +2706,20 @@ namespace Aerospike.Client
 		}
 
 		/// <summary>
+		/// Create PKI user with roles. PKI users are authenticated via TLS and a certificate instead of a password.
+		/// WARNING: This function should only be called for server versions 8.1+. If this function is called for older server versions,
+		/// an error will be returned.
+		/// </summary>
+		/// <param name="policy">admin configuration parameters, pass in null for defaults</param>
+		/// <param name="user">user name</param>
+		/// <param name="roles">variable arguments array of role names.  Predefined roles are listed in Role.cs</param>		
+		public void CreatePKIUser(AdminPolicy policy, string user, IList<string> roles)
+		{
+			AdminCommand command = new();
+			command.CreatePKIUser(cluster, policy, user, roles);
+		}
+
+		/// <summary>
 		/// Remove user from cluster.
 		/// </summary>
 		/// <param name="policy">admin configuration parameters, pass in null for defaults</param>
@@ -2426,15 +2751,29 @@ namespace Aerospike.Client
 
 			AdminCommand command = new AdminCommand();
 
-			if (Util.ByteArrayEquals(userBytes, cluster.user))
+			try
 			{
-				// Change own password.
-				command.ChangePassword(cluster, policy, userBytes, hash);
+				if (Util.ByteArrayEquals(userBytes, cluster.user))
+				{
+					// Change own password.
+					command.ChangePassword(cluster, policy, userBytes, hash);
+				}
+				else
+				{
+					// Change other user's password by user admin.
+					command.SetPassword(cluster, policy, userBytes, hash);
+				}
 			}
-			else
+			catch (AerospikeException ae)
 			{
-				// Change other user's password by user admin.
-				command.SetPassword(cluster, policy, userBytes, hash);
+				if (ae.Result == ResultCode.FORBIDDEN_PASSWORD)
+				{
+					throw new AerospikeException(ResultCode.FORBIDDEN_PASSWORD, "PKI user password not changeable");
+				}
+				else
+				{
+					throw;
+				}
 			}
 			cluster.ChangePassword(userBytes, passwordBytes, hashBytes);
 		}
@@ -2646,7 +2985,7 @@ namespace Aerospike.Client
 
 			try
 			{
-				info = new Info(conn, command);
+				info = new Info(node, conn, command);
 				node.PutConnection(conn);
 			}
 			catch (Exception)
