@@ -1,5 +1,5 @@
 /* 
- * Copyright 2012-2024 Aerospike, Inc.
+ * Copyright 2012-2025 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -24,6 +24,8 @@ namespace Aerospike.Client
 {
 	public sealed class AsyncCluster : Cluster
 	{
+		private readonly AsyncClient asyncClient;
+
 		// Command scheduler.
 		private readonly AsyncScheduler scheduler;
 
@@ -39,9 +41,10 @@ namespace Aerospike.Client
 		// Maximum async connections per node.
 		internal readonly int asyncMaxConnsPerNode;
 
-		public AsyncCluster(AsyncClientPolicy policy, Host[] hosts)
-			: base(policy, hosts)
+		public AsyncCluster(AsyncClient client, AsyncClientPolicy policy, string configPath, Host[] hosts)
+			: base(client, policy, configPath, hosts)
 		{
+			this.asyncClient = client;
 			maxCommands = policy.asyncMaxCommands;
 			asyncMinConnsPerNode = policy.asyncMinConnsPerNode;
 			asyncMaxConnsPerNode = (policy.asyncMaxConnsPerNode >= 0) ? policy.asyncMaxConnsPerNode : policy.maxConnsPerNode;
@@ -71,6 +74,7 @@ namespace Aerospike.Client
 					throw new AerospikeException(ResultCode.PARAMETER_ERROR, "Unsupported MaxCommandAction value: " + policy.asyncMaxCommandAction.ToString());
 			}
 
+			UpdateClusterConfig(true);
 			StartTendThread(policy);
 		}
 
@@ -93,6 +97,17 @@ namespace Aerospike.Client
 		public void ReleaseBuffer(BufferSegment segment)
 		{
 			scheduler.Release(segment);
+		}
+
+		protected override AsyncClient GetAerospikeClient()
+		{
+			return asyncClient;
+		}
+
+		protected override void UpdateClientPolicy()
+		{
+			var client = GetAerospikeClient();
+			client.clientPolicy = new AsyncClientPolicy((AsyncClientPolicy)client.GetClientPolicy(), client.configProvider);
 		}
 	}
 }
