@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2023 Aerospike, Inc.
+ * Copyright 2012-2025 Aerospike, Inc.
  *
  * Portions may be licensed to Aerospike, Inc. under one or more contributor
  * license agreements.
@@ -92,7 +92,18 @@ namespace Aerospike.Client
 		/// </summary>
 		public void Write(Command cmd)
 		{
-			cmd.WriteExpHeader(bytes.Length);
+			cmd.WriteExpHeader(bytes.Length, FieldType.FILTER_EXP);
+			Array.Copy(bytes, 0, cmd.dataBuffer, cmd.dataOffset, bytes.Length);
+			cmd.dataOffset += bytes.Length;
+		}
+
+		/// <summary>
+		/// Write expression in wire protocol.
+		/// For internal use only.
+		/// </summary>
+		internal void WriteIndex(Command cmd)
+		{
+			cmd.WriteExpHeader(bytes.Length, FieldType.INDEX_EXPRESSION);
 			Array.Copy(bytes, 0, cmd.dataBuffer, cmd.dataOffset, bytes.Length);
 			cmd.dataOffset += bytes.Length;
 		}
